infrastructure/requirements.txt
Arpeggio==2.0.2
attrs==23.1.0
dill==0.3.7
grpcio==1.56.2
parver==0.4
protobuf==4.24.3

black==23.9.1
isort==5.12.0
flake8==6.1.0

pulumi==3.86.0
pulumi-docker==4.4.1
pulumi-gcp==6.67.0
pulumi-random==4.14.0

PyYAML==6.0.1

semver==2.13.0
six==1.16.0
typing-extensions==4.8.0
